requirements.txt
playwright>=1.41.0
playwright-stealth>=1.0.6
supabase>=2.3.4
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
